chore(layers): move enums and types from utils
