+++ src/types.ts
@@ -1,5 +1,32 @@
 import type { OrientedBoundingBox, BoundingSphere } from "@math.gl/culling";
-import { Theme } from "./enums";
+
+export enum Theme {
+  Dark,
+  Light,
+}
+
+export enum Layout {
+  Default = 'default',
+  Tablet = 'tablet',
+  Mobile = 'mobile'
+}
+
+export enum ComparisonMode {
+  acrossLayers,
+  withinLayer
+}
+
+export enum ListItemType {
+  Radio,
+  Checkbox,
+}
+
+export enum ActiveButton {
+  options,
+  settings,
+  memory,
+  none,
+}
 
 export type TileWarning = {
   type: string;
