--- src/utils/enums.ts
@@ -1,28 +1,0 @@
-
-export enum Theme {
-  Dark,
-  Light,
-}
-
-export enum Layout {
-  Default = 'default',
-  Tablet = 'tablet',
-  Mobile = 'mobile'
-}
-
-export enum ComparisonMode {
-  acrossLayers,
-  withinLayer
-}
-
-export enum ListItemType {
-  Radio,
-  Checkbox,
-}
-
-export enum ActiveButton {
-  options,
-  settings,
-  memory,
-  none,
-}
