+++ src/utils/validation-utils/tile-validation/lod-validation.ts
@@ -1,5 +1,5 @@
 import type { Tile3D } from "@loaders.gl/tiles";
-import type { TileWarning } from '../../types';
+import type { TileWarning } from '../../../types';
 
 import { LOD_WARNING_TYPE, PARENT_LOD_WARNING_TYPE } from "../../../constants/map-styles";
